cat_or_dog/requirements.txt
backcall==0.1.0
cycler==0.10.0
decorator==4.4.2
future==0.18.2
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
opencv-python==4.2.0.34
pandas==1.0.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
powerline-gitstatus==1.3.1
powerline-status==2.7
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
torch==1.5.1
torchvision==0.6.1
tqdm==4.47.0
traitlets==4.3.3
wcwidth==0.2.3
